# services/openapi.yaml
openapi: 3.0.3
info:
  title: SME Cloud-Native Microservices API - Topic 10
  version: 1.0.0
  description: API contract for gateway routing, authorization enforcement, security testing, and API fuzzing.
servers:
  - url: https://localhost:8443
    description: Kong API Gateway HTTPS endpoint
  - url: http://localhost:8000
    description: Kong API Gateway HTTP endpoint for local testing
...
paths:
...
  /api/v1/billing/checkout:
    post:
      tags:
        - Billing
      operationId: createCheckoutSession
      summary: Create a checkout request and trigger payment processing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '202':
          description: Checkout request accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutAcceptedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/webhooks/payment:
    post:
      tags:
        - Webhooks
      operationId: receivePaymentWebhook
      summary: Receive signed payment webhook callback
      parameters:
        - name: X-Webhook-Timestamp
          in: header
          required: true
          schema:
            type: string
          example: "1718272800"
        - name: X-Webhook-Nonce
          in: header
          required: true
          schema:
            type: string
          example: "nonce-123456"
        - name: X-Webhook-Signature
          in: header
          required: true
          schema:
            type: string
          example: "sha256=0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentWebhookRequest'
      responses:
        '200':
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAcceptedResponse'
        '401':
          description: Missing or invalid HMAC signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Expired timestamp or replayed nonce
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Missing, expired, revoked, or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated principal does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CheckoutRequest:
      type: object
      required:
        - order_id
        - amount
      properties:
        order_id:
          type: string
          example: "order-001"
        amount:
          type: number
          format: float
          example: 150000
        currency:
          type: string
          example: "VND"
...
    CheckoutAcceptedResponse:
      type: object
      required:
        - checkout_id
        - status
      properties:
        checkout_id:
          type: string
          example: "checkout-001"
        status:
          type: string
          example: "accepted"
...
    PaymentWebhookRequest:
      type: object
      required:
        - event_id
        - event_type
        - checkout_id
      properties:
        event_id:
          type: string
          example: "evt-001"
        event_type:
          type: string
          example: "payment.succeeded"
        checkout_id:
          type: string
          example: "checkout-001"
        amount:
          type: number
          format: float
          example: 150000
    WebhookAcceptedResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: "accepted"
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "forbidden"
        message:
          type: string
          example: "Access denied"
        correlation_id:
          type: string
          example: "req-123456"